+++ docs/versionhistory.rst
@@ -5,8 +5,8 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 
 **3.1.0**
 
+- Allow parameterless unmarshaller callbacks that return a new instance of the target class
 - Switched YAML serializer to use ruamel.yaml instead of PyYAML
-- Upgraded Asphalt core minimum version to v2.1
 
 **3.0.0** (2016-07-03)